moved reusable functions in testUtils
